Support command arguments as input string.

+++ test.sh
@@ -10,12 +10,21 @@ testsame () {
 actual=`echo 'ÌÚÅá' | $PWD/uimsh-tutctool.scm bushuconv`
 testsame "ÎÂ" "$actual"
 
-ln -s uimsh-tutctool.scm bushuconv
+if [ ! -e bushuconv ]; then
+	ln -s uimsh-tutctool.scm bushuconv
+fi
 actual=`echo 'ÌÚÅá' | $PWD/bushuconv`
 testsame "ÎÂ" "$actual"
 
 actual=`echo '¢¥¢¥ÌÚ¢¥¿Í¿Í¾òÉ×' | $PWD/uimsh-tutctool.scm bushuconv`
 testsame "óÏ" "$actual"
+
+actual=`$PWD/uimsh-tutctool.scm bushuconv 'ÌÚÅá' '¢¥¢¥ÌÚ¢¥¿Í¿Í¾òÉ×'`
+expect='ÎÂ
+óÏ'
+testsame "$expect" "$actual"
+actual=`$PWD/bushuconv 'ÌÚÅá' '¢¥¢¥ÌÚ¢¥¿Í¿Í¾òÉ×'`
+testsame "$expect" "$actual"
 
 actual=`echo 'ÌÚÅá' | $PWD/uimsh-tutctool.scm bushucand`
 testsame "ÎÂÛ·ÛÃÛëÜ¸äíÛÊà®äÌ" "$actual"
@@ -23,9 +32,18 @@ testsame "ÎÂÛ·ÛÃÛëÜ¸äíÛÊà®äÌ" "$actual"
 actual=`echo '¸ıÌÚ¥¤' | $PWD/uimsh-tutctool.scm bushucand`
 testsame "ÊİË«ÔÈèŞêğ" "$actual"
 
-ln -s uimsh-tutctool.scm bushucand
+if [ ! -e bushucand ]; then
+	ln -s uimsh-tutctool.scm bushucand
+fi
 actual=`echo '¸ıÌÚ¥¤' | $PWD/bushucand`
 testsame "ÊİË«ÔÈèŞêğ" "$actual"
+
+actual=`$PWD/uimsh-tutctool.scm bushucand 'ÌÚÅá' '¸ıÌÚ¥¤'`
+expect='ÎÂÛ·ÛÃÛëÜ¸äíÛÊà®äÌ
+ÊİË«ÔÈèŞêğ'
+testsame "$expect" "$actual"
+actual=`$PWD/bushucand 'ÌÚÅá' '¸ıÌÚ¥¤'`
+testsame "$expect" "$actual"
 
 actual=`echo 'Ä·ÎÂ'|$PWD/uimsh-tutctool.scm tutchelp`
 expect="  |  |  |  |  ||  |     |  |           |  ||
@@ -33,9 +51,15 @@ expect="  |  |  |  |  ||  |     |  |           |  ||
   |  |  | d|  ||  |     |  |a(ÎÂ¢¥ÌÚÅá)|  ||
   |  |  |  | e||  |1(Ä·)| f|           |  ||"
 testsame "$expect" "$actual"
+actual=`$PWD/uimsh-tutctool.scm tutchelp 'Ä·ÎÂ'`
+testsame "$expect" "$actual"
 
-ln -s uimsh-tutctool.scm tutchelp
+if [ ! -e tutchelp ]; then
+	ln -s uimsh-tutctool.scm tutchelp
+fi
 actual=`echo 'Ä·ÎÂ'|$PWD/tutchelp`
+testsame "$expect" "$actual"
+actual=`$PWD/tutchelp 'Ä·ÎÂ'`
 testsame "$expect" "$actual"
 
 actual=`echo 'if.g'|$PWD/uimsh-tutctool.scm seq2kanji`
@@ -44,22 +68,45 @@ testsame "Ãæ¸Å" "$actual"
 actual=`echo 'aljdljdjru fjxiala/.;f' | uim-sh $PWD/uimsh-tutctool.scm seq2kanji`
 testsame "ÅÜŞ¹¤Î¿ÊÄ½" "$actual"
 
-ln -s uimsh-tutctool.scm seq2kanji
+if [ ! -e seq2kanji ]; then
+	ln -s uimsh-tutctool.scm seq2kanji
+fi
 actual=`echo 'if.g'|$PWD/seq2kanji`
 testsame "Ãæ¸Å" "$actual"
+
+actual=`$PWD/uimsh-tutctool.scm seq2kanji 'if.g' 'aljdljdjru fjxiala/.;f'`
+expect='Ãæ¸Å
+ÅÜŞ¹¤Î¿ÊÄ½'
+testsame "$expect" "$actual"
+actual=`$PWD/seq2kanji 'if.g' 'aljdljdjru fjxiala/.;f'`
+testsame "$expect" "$actual"
 
 actual=`echo 'ÅÅÃÏµë·îÊ¬Æ°ÅÄ¿·Æ± ' | $PWD/uimsh-tutctool.scm kanji2seq \
 | cut -b 2- | $PWD/uimsh-tutctool.scm seq2kanji`
 testsame "¤¦¤«¤â¤·¤ì¤Ş¤»¤ó¡£" "$actual"
 
-ln -s uimsh-tutctool.scm kanji2seq
+if [ ! -e kanji2seq ]; then
+	ln -s uimsh-tutctool.scm kanji2seq
+fi
 actual=`echo 'ÅÅÃÏµë·îÊ¬Æ°ÅÄ¿·Æ± ' | $PWD/kanji2seq \
 | cut -b 2- | $PWD/seq2kanji`
 testsame "¤¦¤«¤â¤·¤ì¤Ş¤»¤ó¡£" "$actual"
 
+actual=`$PWD/uimsh-tutctool.scm kanji2seq 'ÅÅÃÏµë·îÊ¬Æ°ÅÄ¿·Æ± '`
+testsame "jruekwjsighwkshflf " "$actual"
+actual=`$PWD/kanji2seq 'ÅÅÃÏµë·îÊ¬Æ°ÅÄ¿·Æ± '`
+testsame "jruekwjsighwkshflf " "$actual"
+
 actual=`echo U+25b3 | uim-sh $PWD/uimsh-tutctool.scm kcodeucs`
 testsame "¢¤" "$actual"
 
-ln -s uimsh-tutctool.scm kcodeucs
+if [ ! -e kcodeucs ]; then
+	ln -s uimsh-tutctool.scm kcodeucs
+fi
 actual=`echo U+25b3 | $PWD/kcodeucs`
+testsame "¢¤" "$actual"
+
+actual=`uim-sh $PWD/uimsh-tutctool.scm kcodeucs U+25b3`
+testsame "¢¤" "$actual"
+actual=`uim-sh $PWD/kcodeucs U+25b3`
 testsame "¢¤" "$actual"
